refactor(frontend): move modern track to src/features/ following Vite conventions
Modern track code no longer lives under the legacy static/js/ path.
New features go in src/features/<name>/, separate from legacy code
in src/static/js/. Updated imports, ESLint config, and docs.
