+++ CONTRIBUTING.md
@@ -15,7 +15,7 @@ Bug fixes and minor changes to existing features should follow legacy patterns.
 
 ### Modern Track (new features)
 
-- **Components**: `frontend/src/static/js/features/<feature-name>/`
+- **Components**: `frontend/src/features/<feature-name>/`
 - **State management**: Zustand stores (`useFooStore.js` — `use` prefix, hook-based)
 - **Server state**: TanStack Query (`useQuery`, `useMutation`)
 - **Styling**: Tailwind CSS utilities (import `tailwind.css` — no new SCSS files)
@@ -42,10 +42,10 @@ All new features must use modern track patterns.
 
 ### New Feature Checklist
 
-1. Create a directory under `features/<feature-name>/`
+1. Create a directory under `src/features/<feature-name>/`
 2. Create a Zustand store if client state is needed (`useFooStore.js`)
 3. Use `useQuery`/`useMutation` for API calls (wrap in a scoped `QueryClientProvider`)
-4. Style with Tailwind utilities — import `../../../css/tailwind.css`
+4. Style with Tailwind utilities — import `../../static/css/tailwind.css`
 5. Create an entry file in `src/entries/<feature-name>.js`
 6. Add the entry to `rollupOptions.input` in `vite.config.js`
 7. Create a Django template extending `base.html`
