+++ frontend/src/entries/modern-demo.js
@@ -1,4 +1,4 @@
 import { renderPage } from '../static/js/_helpers.js';
-import ModernDemoPage from '../static/js/features/modern-demo/ModernDemoPage';
+import ModernDemoPage from '../features/modern-demo/ModernDemoPage';
 
 renderPage('page-modern-demo', ModernDemoPage);
